refactor: 정적 메서드 삭제, 한글 입력 추가
