+++ src/test/java/racingcar/view/InputViewTest.java
@@ -13,15 +13,16 @@ class InputViewTest {
     public void setUp() {
         inputView = new InputView();
     }
+
     @Test
-    public void 자동차_이름_입력_예외_테스트() {
+    public void 자동차_이름_검증_예외_테스트() {
         String emptyInput = "";
         assertThatThrownBy(() -> inputView.validateCarNames(emptyInput))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
-    public void 횟수_입력_예외_테스트() {
+    public void 횟수_검증_예외_테스트() {
         String numericInput = "d";
         assertThatThrownBy(() -> inputView.validateAttempts(numericInput))
                 .isInstanceOf(IllegalArgumentException.class);
